Retry single http requests instead of starting all of the work over.
It is very common for a transient 504 error to cause a lot of work to be discarded. This allows less work to be discarded when that happens. This code does it up to 4 times waiting 7 sec in between. This might need to be tuned but from testing usually the next request after a 504 succeeds.